Some more work on a Javascript-based frequency analysis attack.
In theory the process should be okay, but I have a memory leak somewhere that needs plugging. If left to run the program balloons up to silly amounts of memory and dies.

Working with node on Windows seems to be a little difficult, so I may have to set up on Linux and continue from there.